OZ-658: Add config files and structure the project
