+++ openelis-openmrs/src/main/java/com/ozonehis/eip/openelis/openmrs/config/OpenmrsOpenelisConfig.java
@@ -5,15 +5,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.ozonehis.eip.openmrs.openelis.config;
+package com.ozonehis.eip.openelis.openmrs.config;
 
-import org.openmrs.eip.app.config.AppConfig;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.PropertySource;
 
-/**
- * Import the {@link AppConfig} class to ensure that the required beans are created.
- */
 @Configuration
-@Import({AppConfig.class})
-public class OpenmrsWatcherConfig {}
+@PropertySource(
+        value = {"eip-openelis-openmrs.properties", "classpath:eip-openelis-openmrs.properties"},
+        ignoreResourceNotFound = true)
+public class OpenmrsOpenelisConfig {}
