+++ openelis-openmrs/src/main/java/com/ozonehis/eip/openelis/openmrs/config/OpenmrsWatcherConfig.java
@@ -5,13 +5,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.ozonehis.eip.openmrs.openelis.config;
+package com.ozonehis.eip.openelis.openmrs.config;
 
+import org.openmrs.eip.app.config.AppConfig;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.annotation.Import;
 
+/**
+ * Import the {@link AppConfig} class to ensure that the required beans are created.
+ */
 @Configuration
-@PropertySource(
-        value = {"eip-openelis-openmrs.properties", "classpath:eip-openelis-openmrs.properties"},
-        ignoreResourceNotFound = true)
-public class OpenmrsOpenelisConfig {}
+@Import({AppConfig.class})
+public class OpenmrsWatcherConfig {}
