+++ openelis-openmrs/src/main/java/com/ozonehis/eip/openelis/openmrs/handlers/PatientHandler.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.ozonehis.eip.openmrs.openelis.handlers;
+package com.ozonehis.eip.openelis.openmrs.handlers;
 
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
