+++ openelis-openmrs/src/main/java/com/ozonehis/eip/openelis/openmrs/processors/PatientProcessor.java
@@ -5,9 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.ozonehis.eip.openmrs.openelis.processors;
+package com.ozonehis.eip.openelis.openmrs.processors;
 
-import com.ozonehis.eip.openmrs.openelis.handlers.PatientHandler;
+import com.ozonehis.eip.openelis.openmrs.handlers.PatientHandler;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.camel.CamelExecutionException;
