+++ openelis-openmrs/src/main/java/com/ozonehis/eip/openelis/openmrs/routes/OpenmrsWatcherInitRoute.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.ozonehis.eip.openmrs.openelis.routes;
+package com.ozonehis.eip.openelis.openmrs.routes;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.springframework.stereotype.Component;
